requirements.txt
functions-framework==3.*
google-cloud-secret-manager==2.*
google-cloud-scheduler==2.*
